move rest of test files to folders
