+++ frontend/src/tests/job_test/Job.test.js
@@ -1,5 +1,5 @@
 
-import { Job } from "../components";
+import { Job } from "../../components";
 
 test('Job', () => {
     const component = shallow(<Job />);
