+++ frontend/src/tests/job_test/JobList.test.js
@@ -1,4 +1,4 @@
-import { JobList } from "../components";
+import { JobList } from "../../components";
 
 test('JobList', () => {
     const component = shallow(<JobList />);
