+++ frontend/src/tests/job_test/TagView.test.js
@@ -1,4 +1,4 @@
-import { TagView } from '../components';
+import { TagView } from '../../components';
 
 test('TagView', () => {
     const component = shallow(<TagView tag={{}} />);
